# Scale K block size with inverse element size